Start this repo, mostly copy from an archived indicator_configuration repo
The
[indicator_configuration](https://github.com/yonatank93/indicator_configuration)
repo was the original repo where this project started. However, along
the development, I made a mistake of commmitting a large file, and I
don't know how to fix it yet. So, I just started this repo to redeem
that mistake.

For this first commit, I exclude the examples, which have the large
files. I will slowly add those.