UPDATE PWS examples, especially for the singular cases
Previously for the singular cases, I removed the column of config
Jacobian corresponding to buses outside the area of interest and the
rows and columns of the target FIM with zero diagonal elements were
also removed.

In this update, I keep the zero diagonal elements of the target FIM
and keep all columns of config Jacobians. With this formulation,
we assume that the buses still make some measurement of the quantities
corresponding to buses outside the area. From this observation, then
we ask about complete observability of the buses inside the area
only. From the end result, we can potentially still observe the other
buses outside the area.
Whereas in the previous version, we assume that the buses cannot make
observation of the buses outside the area. That's why we removed
some columns in the Jacobians.

I think the new formulation makes more sense, because it is more
practical to use the existing network and observe part of it, without
removing any branches.