EXAMPLES: Update the SW MoS2 examples
It seems that this update gives a slight different optimal configurations
and weights compared to before I made a change in the convex optimization
module. The difference seem to come from some numerical noise in
some of the operations. I am still ok with the new results, since
we still get the same number of optimal configurations, but the
largest weight correspond to requiring force calculation accuracy
within 0.5 meV/atom instead.