Add MISDP version of the information-matching calculation
The MISDP version can give smaller number of optimal PMUs, and the
result is consistent. I then use the MISDP result as an initial guess
to the SDP version to get the non-integer weights.

For the 39-bus case, other people argue that we need at least 8
buses. We are able to achieve this number throught the MISDP version
of the calculation. Unfortunately, the MISDP version cannot identify
minimal buses for the subset version of this case.