Weight upper bound (#6)
* Add an optional upper bound constraint for the weights.

* This is a suggestion from Vasily, based on some problems in
materials science
* I also change the format to the solver input argument to be a
keyword arguments. I think this format is better.
* And I separate the tolerance to get the nonzero weights into the
primer and dual tolerances.

* Add a test for the weight upper bound

* Update the examples

* Make so that user can do `from information_matching import ConvexOpt`

The `ConvexOpt` module is the main module of this package. The
information-matching calculation to find the optimal weights is in
this module.
So, it is natural to make it easier for the user to import this module.

* Bump up version to 1.1.0

* This package also works for python 3.10