Refactoring (#8)
* Add a function to help precondition the FIMs

The preconditioning function will compute the scaling factor, so that
it is easier for the user to add scaling factor to the input FIMs for
better convergence. Currently, there are only 2 type of
preconditioning or scaling factor:

1. Frobenius norm for each FIM
2. Uniform scaling factor using the largest Frobenius norm

* Update the examples to use preconditioning function and some clean up

* When input fim is an array, I added an if statement

This  if statement checks if the scale_type is frobenius type,
which will give the same answer. But in the future, we might add
other kind of preconditioning, so I added the if statement to
anticipate this update in the future.

* Add parameter transformation module

Currently, there are only affine and logarithmic transformation.

* Convert the documentation to numpy style.

* Add CombinedTransform class that can be used to combine multiple transformations

* Change documentation format in finitediff.py to numpydoc style

This is for consistency since the rest of documentation uses numpydoc
style too.

* Rename mcmc_utils.py -> sampling_utils.py and some clean up

I think sampling_utils.py is a more appropriate name, because this
package is not about MCMC. However, we can still utilize some sampling
analysis techniques in MCMC to do some analysis here. So, I still want
to include these functions andd remove other functions that are not
relevant, like the prior.

* Reorganing utility functions

I just realized that there are quite a lot of utility functions, like
those in misc.py (previously named utils.py), termination.py, etc. I
don't want these utility functions to obstruct the main
functions/modules. So, I collected them and put them into the same
utils folder.

I made necessary updates to this change.

* Update the examples with the reorganization of utils

* Rename module preconditioning->precondition and function preconditioning->preconditioner

* I forgot to add transform and precondition information in README

* Add a function that wraps over input function and does parameter transformation

This function essentially apply inverse_transform and calls the input
function. The utility of this function is to be used to take
derivative of the input function in transformed parameter space.

* Change AffineTransform to a more general form, separating the shifts

With this more general form, this AffineTransform class can also be
used to fix some parameters.

* Update documentation and add additional notes to help users

* Fix attribute error in test_leastsq.py due to addition of "_" in some functions' names

* Resolve lint issue because test_affine_transform is renamed to test_transform

* Bump up version 1.1.0 -> 1.2.0

* Remove CI for Python 3.7 and extend CI to Python 3.11 and 3.12