Add more transformations (#9)
* Rename CombinedTransform -> SplitTransform and add ComposedTransform

These 2 classes are for applying multiple transformations, but they
work differently. SplitTransform applies each element of
transformations list argument to a subset of parameters. While
ComposedTransform compose the transformations and apply them to all parameters.

* Add AtanhTransform that can be used when parameters are bounded between some range.

* Update the test in the following ways:

* Add test for ComposedTransform and AtanhTransform.
* Add a test that check if some transformations satisfy their constraints.

* Change keys in avail_transform dictionary

The keys now are just the name of the transformation class. This will
make it easier to deciede on the name, and for other applications.